@@ -4,8 +4,8 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  flutter: ">=3.10.0"
-  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.24.0"
+  sdk: ">=3.5.0 <4.0.0"
 
 dependencies:
   # The following adds the Cupertino Icons font to your application.
@@ -16,16 +16,11 @@ dependencies:
   medea_flutter_webrtc:
     path: ../
 
-dependency_overrides:
-  # Pinned until `gql-dart` updates to 4 version:
-  # https://github.com/gql-dart/gql/blob/master/links/gql_websocket_link/pubspec.yaml#L13
-  uuid: ^3.0.7
-
 dev_dependencies:
-  build_runner: ^2.1.10
+  build_runner: ^2.4.11
   flutter_driver:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
   integration_test:
