@@ -6,33 +6,28 @@ version: 0.10.0
 homepage: https://github.com/instrumentisto/medea-flutter-webrtc
 
 environment:
-  flutter: ">=3.10.0"
-  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.24.0"
+  sdk: ">=3.5.0 <4.0.0"
 
 dependencies:
-  collection: ^1.17.0
-  ffi: ^2.0.1
-  js: ^0.6.5
+  collection: ^1.17.2
+  ffi: ^2.0.2
   flutter:
     sdk: flutter
   flutter_rust_bridge: 1.82.4  # should be the same as in `Cargo.lock`
-  freezed_annotation: ^2.4.1
-  meta: ^1.8.0
-  uuid: ^4.2.1
+  freezed_annotation: ^2.4.4
+  uuid: ^4.4.2
   web: ">=0.5.1 <2.0.0"
 dev_dependencies:
-  build_runner: ^2.3.2
+  build_runner: ^2.4.12
   ffigen: 9.0.0
-  flutter_lints: ^3.0.1
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
-  freezed: ^2.4.3
+  freezed: ^2.5.7
   import_sorter: ^4.6.0
 
 dependency_overrides:
-  # Pinned until `gql-dart` updates to 4 version:
-  # https://github.com/gql-dart/gql/blob/master/links/gql_websocket_link/pubspec.yaml#L13
-  uuid: ^3.0.7
   ffigen: 9.0.0
 
 flutter:
