@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:medea_flutter_webrtc/medea_flutter_webrtc.dart';
 
 class OnDeviceChangeNotifierSample extends StatefulWidget {
-  const OnDeviceChangeNotifierSample({Key? key}) : super(key: key);
+  const OnDeviceChangeNotifierSample({super.key});
 
   @override
   State<OnDeviceChangeNotifierSample> createState() => _State();
